requirements.txt
pandas==2.2.3
numpy==1.26.4
pyyaml==6.0
pytest==7.4.2
